app/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
mysql-connector-python==9.0.0
python-multipart==0.0.9
pypdf==4.3.1
beautifulsoup4==4.12.3
requests==2.32.3
passlib[bcrypt]==1.7.4
boto3==1.34.141